+++ template/config/webpack.config.base.js
@@ -1,33 +1,45 @@
-const helpers = require("./helpers"),
+const helpers = require('./helpers'),
   CopyWebpackPlugin = require('copy-webpack-plugin');
 
 let config = {
   entry: {
-    "main": helpers.root("/src/main.ts")
+    'main': helpers.root('/src/main.ts')
   },
   output: {
-    path: helpers.root("/dist/js"),
-    filename: "[name].js"
+    path: helpers.root('/dist'),
+    filename: 'js/[name].[hash].js'
   },
-  devtool: "source-map",
+  devtool: 'source-map',
   resolve: {
-    extensions: [".ts", ".js", ".html"],
+    extensions: ['.ts', '.js', '.html'],
     alias: {
-      'vue$': 'vue/dist/vue.common.js'
+      'vue$': 'vue/dist/vue.esm.js',
     }
   },
   module: {
-    rules: [
-      {test: /\.ts$/, exclude: /node_modules/, enforce: 'pre', loader: 'tslint-loader'},
-      {test: /\.ts$/, exclude: /node_modules/, loader: "awesome-typescript-loader"},
-      {test: /\.html$/, loader: 'raw-loader', exclude: ['./src/index.html']}
+    rules: [{
+        test: /\.ts$/,
+        exclude: /node_modules/,
+        enforce: 'pre',
+        loader: 'tslint-loader'
+      },
+      {
+        test: /\.ts$/,
+        exclude: /node_modules/,
+        loader: 'awesome-typescript-loader'
+      },
+      {
+        test: /\.html$/,
+        loader: 'raw-loader',
+        exclude: ['./src/index.html']
+      }
     ],
   },
   plugins: [
-    new CopyWebpackPlugin([
-      {from: 'src/assets', to: '../assets'},
-      {from: 'src/css', to: '../css'}
-    ]),
+    new CopyWebpackPlugin([{
+      from: 'src/assets',
+      to: './assets'
+    }, ]),
   ]
 };