+++ template/config/webpack.config.coverage.js
@@ -1,22 +1,12 @@
-const webpackConfig = require("./webpack.config.base")
-  DefinePlugin = require('webpack/lib/DefinePlugin'),
-  SourceMapDevToolPlugin = require('webpack/lib/SourceMapDevToolPlugin'),
-  env = require('../environment/dev.env');
-
-webpackConfig.module.rules[1].query = {
-  compilerOptions: {
-    inlineSourceMap: true,
-    sourceMap: false
-  }
-};
+const webpackConfig = require('./webpack.config.test');
 
 webpackConfig.module.rules = [...webpackConfig.module.rules,
   {
     test: /\.ts$/,
-    enforce: "post",
-    loader: "istanbul-instrumenter-loader",
+    enforce: 'post',
+    loader: 'istanbul-instrumenter-loader',
     exclude: [
-      "node_modules",
+      'node_modules',
       /\.spec\.ts$/
     ],
     query: {
@@ -24,17 +14,5 @@ webpackConfig.module.rules = [...webpackConfig.module.rules,
     }
   }
 ];
-
-webpackConfig.plugins = [...webpackConfig.plugins,
-  new SourceMapDevToolPlugin({
-    filename: null, // if no value is provided the sourcemap is inlined
-    test: /\.(ts|js)($|\?)/i
-  }),
-  new DefinePlugin({
-    'process.env': env
-  })
-];
-
-webpackConfig.devtool = "inline-source-map";
 
 module.exports = webpackConfig;