+++ template/config/webpack.config.dev.js
@@ -1,21 +1,50 @@
-const helpers = require("./helpers"),
-  webpackConfig = require("./webpack.config.base"),
+const helpers = require('./helpers'),
+  webpackConfig = require('./webpack.config.base'),
+  HtmlWebpackPlugin = require('html-webpack-plugin'),
   DefinePlugin = require('webpack/lib/DefinePlugin'),
   env = require('../environment/dev.env');
 
-webpackConfig.devServer = {
-  port: 8080,
-  host: "localhost",
-  historyApiFallback: true,
-  watchOptions: {aggregateTimeout: 300, poll: 1000},
-  contentBase: './src',
-  open: true
-};
+webpackConfig.module.rules = [...webpackConfig.module.rules,
+  {
+    test: /\.scss$/,
+    use: [{
+        loader: 'style-loader'
+      },
+      {
+        loader: 'css-loader'
+      },
+      {
+        loader: 'sass-loader'
+      }
+    ]
+  },
+  {
+    test: /\.(jpg|png|gif|eot|svg|ttf|woff|woff2)$/,
+    loader: 'file-loader'
+  }
+];
 
 webpackConfig.plugins = [...webpackConfig.plugins,
+  new HtmlWebpackPlugin({
+    inject: true,
+    template: helpers.root('/src/index.html'),
+    favicon: helpers.root('/src/favicon.ico')
+  }),
   new DefinePlugin({
     'process.env': env
   })
-]
+];
+
+webpackConfig.devServer = {
+  port: 8080,
+  host: 'localhost',
+  historyApiFallback: true,
+  watchOptions: {
+    aggregateTimeout: 300,
+    poll: 1000
+  },
+  contentBase: './src',
+  open: true
+};
 
 module.exports = webpackConfig;