+++ template/src/components/home/home.ts
@@ -1,6 +1,8 @@
 import Vue from 'vue';
 import Component from 'vue-class-component';
 
+import './home.scss';
+
 @Component({
     template: require('./home.html')
 })