+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Change log
 
+## [v0.9.0] (2017-08-24)
+
+**Features**
+
+* Add [html-webpack-plugin](https://github.com/jantimon/html-webpack-plugin) for creating HTML files
+* Add webpack loaders for loading sass styles
+* Add [favicon webpack generator](https://github.com/jantimon/favicons-webpack-plugin) for creating favicons and icons!
+
 ## [v0.8.0] (2017-08-24)
 
 **Features**
@@ -101,6 +109,7 @@
 [#3]: https://github.com/ducksoupdev/vue-webpack-typescript/pull/3
 [#2]: https://github.com/ducksoupdev/vue-webpack-typescript/pull/2
 [#1]: https://github.com/ducksoupdev/vue-webpack-typescript/pull/1
+[v0.9.0]: https://github.com/ducksoupdev/vue-webpack-typescript/compare/v0.8.0...v0.9.0
 [v0.8.0]: https://github.com/ducksoupdev/vue-webpack-typescript/compare/v0.7.1...v0.8.0
 [v0.7.1]: https://github.com/ducksoupdev/vue-webpack-typescript/compare/v0.7.0...v0.7.1
 [v0.7.0]: https://github.com/ducksoupdev/vue-webpack-typescript/compare/v0.6.0...v0.7.0